backend/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
faiss-cpu
langchain-community
pydantic
PyPDF2langchain-google-community
apscheduler